fix(azure): update intel
